Move Kusto helpers to shared project
In preparation for #3903, where we will want to push data to the Kusto
database from Maestro.Web, this change moves the Kusto helpers to a
shared project that can be referenced by both AzureDevOpsTimeline and
Maestro.Web.
