+++ arcade-services.sln
@@ -61,6 +61,8 @@ Project("{A07B5EB6-E848-4116-A8D0-A826331D98C6}") = "TelemetryApplication", "src
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureDevOpsTimeline", "src\Telemetry\AzureDevOpsTimeline\AzureDevOpsTimeline.csproj", "{4C27D3E3-3F0C-479B-A9AB-2D9EBCE88503}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "src\Shared\Shared.csproj", "{AB5501C3-C238-4F89-8815-81C26EE7DC16}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -395,6 +397,18 @@ Global
 		{4C27D3E3-3F0C-479B-A9AB-2D9EBCE88503}.Release|x64.Build.0 = Release|Any CPU
 		{4C27D3E3-3F0C-479B-A9AB-2D9EBCE88503}.Release|x86.ActiveCfg = Release|Any CPU
 		{4C27D3E3-3F0C-479B-A9AB-2D9EBCE88503}.Release|x86.Build.0 = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|x64.Build.0 = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Debug|x86.Build.0 = Debug|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|x64.ActiveCfg = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|x64.Build.0 = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|x86.ActiveCfg = Release|Any CPU
+		{AB5501C3-C238-4F89-8815-81C26EE7DC16}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
