+++ src/Shared/CsvWriterExtensions.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace Microsoft.DotNet.AzureDevOpsTimeline
+namespace Microsoft.DotNet.Shared
 {
     public static class CsvWriterExtensions
     {
