+++ src/Telemetry/AzureDevOpsTimeline/AzureDevOpsTimeline.cs
@@ -15,6 +15,7 @@ using Kusto.Data.Exceptions;
 using Kusto.Data.Net.Client;
 using Kusto.Ingest;
 using Microsoft.DotNet.ServiceFabric.ServiceHost;
+using Microsoft.DotNet.Shared;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
